+++ test/framework/src/main/java/org/elasticsearch/logsdb/datageneration/fields/LongFieldDataGenerator.java
@@ -6,30 +6,30 @@
  * Side Public License, v 1.
  */
 
-package org.elasticsearch.datastreams.logsdb.datageneration.fields;
+package org.elasticsearch.logsdb.datageneration.fields;
 
 import org.elasticsearch.core.CheckedConsumer;
-import org.elasticsearch.datastreams.logsdb.datageneration.FieldDataGenerator;
+import org.elasticsearch.logsdb.datageneration.FieldDataGenerator;
 import org.elasticsearch.xcontent.XContentBuilder;
 
 import java.io.IOException;
 
-import static org.elasticsearch.test.ESTestCase.randomAlphaOfLengthBetween;
+import static org.elasticsearch.test.ESTestCase.randomLong;
 
-public class KeywordFieldDataGenerator implements FieldDataGenerator {
+public class LongFieldDataGenerator implements FieldDataGenerator {
     private final String fieldName;
 
-    public KeywordFieldDataGenerator(String fieldName) {
+    public LongFieldDataGenerator(String fieldName) {
         this.fieldName = fieldName;
     }
 
     @Override
     public CheckedConsumer<XContentBuilder, IOException> mappingWriter() {
-        return b -> b.startObject(fieldName).field("type", "keyword").endObject();
+        return b -> b.startObject(fieldName).field("type", "long").endObject();
     }
 
     @Override
     public CheckedConsumer<XContentBuilder, IOException> fieldValueGenerator() {
-        return b -> b.field(fieldName, randomAlphaOfLengthBetween(0, 50));
+        return b -> b.field(fieldName, randomLong());
     }
 }
